Started training and getting results

+++ training_functions.py
@@ -7,14 +7,17 @@ from torch import nn
 from torch import optim
 from torch.distributions import Uniform
 
-def run_training_step(model:AnimalDetector, optimiser:optim.Adam, data_handler:ODDataHandler, batch:tuple):
+def calculate_training_losses(device, model:AnimalDetector, data_handler:ODDataHandler, batch:tuple, augment=True):
 
     with torch.enable_grad():
 
-        optimiser.zero_grad()
         model.train()
 
-        x_ims, xc_seqs, xb_seqs, yc, yb = data_handler.build_xy_batch(batch)
+        x_ims, xc_seqs, xb_seqs, yc, yb = data_handler.build_xy_batch(batch, augment=augment)
+
+        x_ims = x_ims.to(device)
+        yc = yc.to(device)
+        yb = yb.to(device)
 
         ypcs, ypbs = model(x_ims, xc_seqs, xb_seqs)
         
@@ -49,19 +52,18 @@ def run_training_step(model:AnimalDetector, optimiser:optim.Adam, data_handler:O
 
         IoU_loss = (1 - A_inter / A_union).sum(dim=0, keepdim=True)
 
-        total_loss = ce_loss + mse_loss + IoU_loss
+        return ce_loss, mse_loss, IoU_loss
 
-        total_loss.backward()
-        optimiser.step()
-
-        return ce_loss.item(), mse_loss.item(), IoU_loss.item(), total_loss.item()
-
-def run_validation_batch(model:AnimalDetector, data_handler:ODDataHandler, batch:tuple):
+def calculate_validation_losses(device, model:AnimalDetector, data_handler:ODDataHandler, batch:tuple):
 
     with torch.no_grad():
 
         model.eval()
-        x_ims, xc_seqs, xb_seqs, yc, yb = data_handler.build_xy_batch(batch)
+        x_ims, xc_seqs, xb_seqs, yc, yb = data_handler.build_xy_batch(batch, augment=False)
+
+        x_ims = x_ims.to(device)
+        yc = yc.to(device)
+        yb = yb.to(device)
 
         ypc, ypb = model(x_ims, xc_seqs, xb_seqs)
 
@@ -105,14 +107,14 @@ def run_validation_batch(model:AnimalDetector, data_handler:ODDataHandler, batch
 
 class HPOptimiser:
 
-    def __init__(self, x_inits, x_bounds, kappa = .1, pars_threshold = 1e-6, x_threshold = 1e-6):
+    def __init__(self, x_inits, x_bounds, kappa = 2.5, pars_threshold = 1e-6):
 
         self.log_length = torch.tensor(0., dtype=torch.float32)
         self.log_sigf = torch.tensor(0., dtype=torch.float32)
         self.log_sign = torch.tensor(0., dtype=torch.float32)
 
         self.kappa = kappa
-        self.x_threshold = x_threshold
+        self.x_threshold = .1 * (x_bounds[1] - x_bounds[0])
         self.pars_threshold = pars_threshold
         self.x_bounds = x_bounds
         
@@ -133,6 +135,7 @@ class HPOptimiser:
         if self.x_obs.size(0) != self.y_obs.size(0):
             return
         
+        y_normed = (self.y_obs - self.y_obs.mean()) / self.y_obs.std()
         with torch.enable_grad():
             curr_pars = [self.log_length.requires_grad_(True), 
                          self.log_sigf.requires_grad_(True),
@@ -156,10 +159,10 @@ class HPOptimiser:
                 
                 L_ky = torch.linalg.cholesky(K_y)
 
-                alpha = torch.linalg.solve(L_ky, self.y_obs)
+                alpha = torch.linalg.solve(L_ky, y_normed)
                 K_y_inv_y =  torch.linalg.solve(L_ky.T, alpha)
 
-                log_marg_likelihood = -.5 * self.y_obs @ K_y_inv_y - torch.log(torch.diag(L_ky)).sum()
+                log_marg_likelihood = -.5 * y_normed @ K_y_inv_y - torch.log(torch.diag(L_ky)).sum()
                 neg_log_marg_likelihood = -log_marg_likelihood
 
                 neg_log_marg_likelihood.backward()
@@ -202,11 +205,12 @@ class HPOptimiser:
         self.log_sigf = state["log sigf"]
         self.log_sign = state["log sign"]
         self.pars_threshold = state["pars threshold"]
-        self.x_threshold = state["x threshold"]
+        self.x_threshold = .1 * (self.x_bounds[1] - self.x_bounds[0])
 
     def select_new_x(self, verbose : bool = False):
 
         lr = .25 * torch.abs(torch.tensor(self.x_bounds[0]) - torch.tensor(self.x_bounds[1]))
+        y_normed = (self.y_obs - self.y_obs.mean()) / self.y_obs.std()
 
         if self.x_obs.size(0) != self.y_obs.size(0):
             return self.x_obs[self.y_obs.size(0)]
@@ -234,7 +238,7 @@ class HPOptimiser:
                 K_x_star_x = sigf * torch.exp(-((self.x_obs - x) ** 2) / (2 * l))
                 K_x_star_x_star = sigf
 
-                alpha = torch.linalg.solve(LK_y, self.y_obs)
+                alpha = torch.linalg.solve(LK_y, y_normed)
                 K_y_inv_y = torch.linalg.solve(LK_y.T, alpha)
 
                 mu_star = K_x_star_x @ K_y_inv_y
@@ -245,7 +249,7 @@ class HPOptimiser:
                 Sig_star = torch.sqrt(K_x_star_x_star - K_x_star_x @ K_y_inv_K_x_star)
 
                 LCB = mu_star - self.kappa * Sig_star
-                (-LCB).backward()
+                LCB.backward()
                 x_optim.step()
 
                 if x < self.x_bounds[0] or x > self.x_bounds[1]:
@@ -272,17 +276,8 @@ class CheckpointSaverLoader:
             self.get_curr_hp()
 
     def get_curr_hp(self):
-        hps = [dir for dir in os.listdir(self.main_dir) if dir != "hpo_state.json"]
-        curr_hp = hps[0]
-        index = int(curr_hp.replace("hp_", ""))
-
-        for hp in hps:
-            n_index = int(hp.replace("hp_", ""))
-            if index < n_index:
-                curr_hp = hp
-                index = n_index
-
-        self.curr_hp = curr_hp
+        self.curr_hp_index = max([int(dir[3:]) for dir in os.listdir(self.main_dir) if dir != "hpo_state.json"])
+        self.curr_hp = "hp_" + str(self.curr_hp_index)
     
     def save_data_handler_state(self, state):
 
@@ -290,15 +285,33 @@ class CheckpointSaverLoader:
         with open(datasets_path, "w") as f:
             json.dump(state, f)
 
+    def load_data_handler_state(self):
+
+        datasets_path = self.main_dir + "//" + self.curr_hp + "//datasets//data.json"
+        with open(datasets_path, "r") as f:
+            state = json.load(f)
+
+        return state
+
     def save_hpo_state(self, state):
 
         with open(self.main_dir + "//hpo_state.json", "w") as f:
             json.dump(state, f)
 
+    def load_hpo_state(self):
+
+        with open(self.main_dir + "//hpo_state.json", "r") as f:
+            state = json.load(f)
+        return state
+
     def save_hp(self, hp:float, hp_index:int):
 
-        self.curr_hp = f"hp_{hp_index}"
+        self.get_curr_hp()
         hp_folder = self.main_dir + f"//hp_{hp_index}"
+
+        if os.path.exists(hp_folder):
+            return
+        
         os.mkdir(hp_folder)
 
         with open(hp_folder + "//hp.txt", "w") as f:
@@ -308,38 +321,57 @@ class CheckpointSaverLoader:
         os.mkdir(hp_folder + "//metrics")
         os.mkdir(hp_folder + "//datasets")
 
-        with open(hp_folder + "//metrics//metrics.json", "w") as f:
-            metrics = {"finished training": False, 
-                       "stopped early": None, 
-                       "num epochs": None, 
-                       "training losses": {"ce loss": [], 
-                                           "mse loss": [], 
-                                           "IoU loss": [],
-                                           "scaled loss":[], 
-                                           "total loss": []},
-                       "validation losses": {"ce loss": [], 
-                                           "mse loss": [], 
-                                           "IoU loss": [],
-                                           "scaled loss":[], 
-                                           "total loss": []}, 
-                       "validation metrics": {"accuracy": [], 
-                                              "mean absolute error":[],
-                                              "mean IoU loss":[]}}
-            
-            json.dump(metrics, f)
+        if not os.path.exists(hp_folder + "//metrics//metrics.json"):
+            with open(hp_folder + "//metrics//metrics.json", "w") as f:
+                metrics = {"finished training": False, 
+                        "stopped early": False,
+                        "early stop threshold":None, 
+                        "num epochs": 0, 
+                        "training losses": {"ce loss": [], 
+                                            "mse loss": [], 
+                                            "IoU loss": [],
+                                            "scaled loss params":[],
+                                            "scaled loss":[], 
+                                            "total loss": []},
+                        "validation losses": {"ce loss": [], 
+                                            "mse loss": [], 
+                                            "IoU loss": [],
+                                            "total loss": []}, 
+                        "validation metrics": {"accuracy": [], 
+                                                "mean absolute error":[],
+                                                "mean IoU loss":[]}}
+                
+                json.dump(metrics, f)
 
-    def save_current_model_optimiser(self, model_state_dict, optim_state_dict):
+    def save_current_model_optimiser(self, model_state_dict, optim_state_dict, loss_scale_params):
 
         model_dir = self.main_dir + "//" + self.curr_hp + "//models"
         torch.save(model_state_dict, model_dir + "//current_model.pth")
         torch.save(optim_state_dict, model_dir + "//optim.pth")
 
+        with open(model_dir + "//loss_scale_params.json", "w") as f:
+            data = {"log_sig_ce":loss_scale_params[0], 
+                    "log_sig_mse":loss_scale_params[1], 
+                    "log_sig_IoU":loss_scale_params[2]}
+            json.dump(data, f)
+
+    def load_current_model_optimiser(self):
+
+        model_dir = self.main_dir + "//" + self.curr_hp + "//models"
+        current_model_state = torch.load(model_dir + "//current_model.pth")
+        current_optimiser_state = torch.load(model_dir + "//optim.pth")
+
+        with open(model_dir + "//loss_scale_params.json", "r") as f:
+            loss_scale_params = json.load(f)
+
+        return current_model_state, current_optimiser_state, loss_scale_params
+     
     def save_best_model(self, model_state_dict):
 
         model_dir = self.main_dir + "//" + self.curr_hp + "//models"
         torch.save(model_state_dict, model_dir + "//best_model.pth")
 
-    def save_training_losses(self, ce_loss, mse_loss, IoU_loss, scaled_loss, total_loss):
+    def save_training_losses(self, ce_loss, mse_loss, IoU_loss, scaled_loss_params, scaled_loss, total_loss):
 
         metrics_path = self.main_dir + "//" + self.curr_hp + "//metrics//metrics.json"
         with open(metrics_path, "r") as f:
@@ -348,13 +380,73 @@ class CheckpointSaverLoader:
         metrics["training losses"]["ce loss"].append(ce_loss)
         metrics["training losses"]["mse loss"].append(mse_loss)
         metrics["training losses"]["IoU loss"].append(IoU_loss)
+        metrics["training losses"]["scaled loss params"].append(scaled_loss_params)
         metrics["training losses"]["scaled loss"].append(scaled_loss)
         metrics["training losses"]["total loss"].append(total_loss)
 
         with open(metrics_path, "w") as f:
             json.dump(metrics, f)
 
-    def save_validation_losses(self, ce_loss, mse_loss, IoU_loss, scaled_loss, total_loss):
+    def save_epoch(self, epoch):
+
+        metrics_path = self.main_dir + "//" + self.curr_hp + "//metrics//metrics.json"
+        with open(metrics_path, "r") as f:
+            metrics = json.load(f)
+
+        metrics["num epochs"] = epoch
+
+        with open(metrics_path, "w") as f:
+            json.dump(metrics, f)
+
+    def load_epoch(self):
+
+        metrics_path = self.main_dir + "//" + self.curr_hp + "//metrics//metrics.json"
+        with open(metrics_path, "r") as f:
+            metrics = json.load(f)
+
+        return metrics["num epochs"]
+    
+    def save_early_stopping(self):
+
+        metrics_path = self.main_dir + "//" + self.curr_hp + "//metrics//metrics.json"
+        with open(metrics_path, "r") as f:
+            metrics = json.load(f)
+
+        metrics["stopped early"] = True
+        with open(metrics_path, "w") as f:
+            json.dump(metrics, f)
+
+    def save_early_stop_threshold(self, threshold):
+
+        metrics_path = self.main_dir + "//" + self.curr_hp + "//metrics//metrics.json"
+        with open(metrics_path, "r") as f:
+            metrics = json.load(f)
+
+        metrics['early stop threshold'] = threshold
+
+        with open(metrics_path, "w") as f:
+            json.dump(metrics, f)
+
+    def load_early_threshold(self):
+
+        metrics_path = self.main_dir + "//" + self.curr_hp + "//metrics//metrics.json"
+        with open(metrics_path, "r") as f:
+            metrics = json.load(f)
+
+        threshold = metrics['early stop threshold']
+        return threshold
+
+    def save_finished_training(self):
+
+        metrics_path = self.main_dir + "//" + self.curr_hp + "//metrics//metrics.json"
+        with open(metrics_path, "r") as f:
+            metrics = json.load(f)
+
+        metrics["finished training"] = True
+        with open(metrics_path, "w") as f:
+            json.dump(metrics, f)
+
+    def save_validation_losses(self, ce_loss, mse_loss, IoU_loss, total_loss):
 
         metrics_path = self.main_dir + "//" + self.curr_hp + "//metrics//metrics.json"
         with open(metrics_path, "r") as f:
@@ -363,7 +455,6 @@ class CheckpointSaverLoader:
         metrics["validation losses"]["ce loss"].append(ce_loss)
         metrics["validation losses"]["mse loss"].append(mse_loss)
         metrics["validation losses"]["IoU loss"].append(IoU_loss)
-        metrics["validation losses"]["scaled loss"].append(scaled_loss)
         metrics["validation losses"]["total loss"].append(total_loss)
 
         with open(metrics_path, "w") as f:
@@ -379,6 +470,32 @@ class CheckpointSaverLoader:
         metrics["validation metrics"]["mean absolute error"].append(mae)
         metrics["validation metrics"]["mean IoU loss"].append(IoU)
 
+        with open(metrics_path, "w") as f:
+            json.dump(metrics, f)
+
     def load_checkpoint(self):
 
-        pass
+        hp_folder = self.main_dir + "//" + self.curr_hp
+
+        metrics_path = hp_folder + "//metrics//metrics.json"
+        with open(metrics_path, "r") as f:
+            metrics = json.load(f)
+
+        assert(metrics["finished training"] is False)
+        num_epochs = self.load_epoch()
+        best_vloss = min(metrics["validation losses"]["total loss"])
+        hpo_state = self.load_hpo_state()
+
+        current_model_state_dict, current_optimiser_state_dict, loss_scale_params = self.load_current_model_optimiser()
+        dh_state = self.load_data_handler_state()
+
+        early_stop_threshold = self.load_early_threshold()
+
+        return {"current model state" : current_model_state_dict,
+                "current optimiser state" : current_optimiser_state_dict,
+                "early stop threshold" : early_stop_threshold, 
+                "num epochs" : num_epochs, 
+                "best vloss" : best_vloss, 
+                "hpo state" : hpo_state, 
+                "loss scale params": loss_scale_params, 
+                "dh state" : dh_state}
